+++ src-docs/src/routes.js
@@ -219,6 +219,9 @@ import { ResponsiveExample }
 import { SearchBarExample }
   from './views/search_bar/search_bar_example';
 
+import { SelectableExample }
+  from './views/selectable/selectable_example';
+
 import { SideNavExample }
   from './views/side_nav/side_nav_example';
 
@@ -300,7 +303,7 @@ const createExample = (example) => {
     throw new Error(`One of your example pages is undefined. This usually happens when you export or import it with the wrong name.`);
   }
 
-  const { title, intro, sections } = example;
+  const { title, intro, sections, beta } = example;
   sections.forEach(section => {
     section.id = slugify(section.title || title);
   });
@@ -312,7 +315,7 @@ const createExample = (example) => {
 
   const component = () => (
     <EuiErrorBoundary>
-      <GuidePage title={title} intro={intro}>
+      <GuidePage title={title} intro={intro} isBeta={beta}>
         {renderedSections}
       </GuidePage>
     </EuiErrorBoundary>
@@ -418,6 +421,7 @@ const navigation = [{
     FilterGroupExample,
     RangeControlExample,
     SearchBarExample,
+    SelectableExample,
   ].map(example => createExample(example)),
 },
 {